test: added tests for search suggestions
