+++ packages/world_countries/test/src/extensions/build_context_extension_test.dart
@@ -31,90 +31,6 @@ void main() => group("BuildContextExtension", () {
         },
       );
 
-      testWidgets("navigator", (tester) async {
-        final context = await contextExtractor(tester);
-        expect(context.navigator, isA<NavigatorState>());
-        expect(context.navigator.canPop(), isFalse);
-      });
-
-      testWidgets("pop pops the current route", (tester) async {
-        final route = MaterialPageRoute<Object>(builder: (_) => Container());
-        await tester.pumpMaterialApp(
-          Builder(
-            builder: (context) => ElevatedButton(
-              onPressed: () async => context.push(route),
-              child: const Text("Push route"),
-            ),
-          ),
-        );
-        await tester.tap(find.byType(ElevatedButton));
-        await tester.pumpAndSettle();
-        final context = tester.element(find.byType(Container));
-        expect(find.byType(Container), findsOneWidget);
-        context.pop();
-        await tester.pumpAndSettle();
-        expect(find.byType(Container), findsNothing);
-      });
-
-      testWidgets("push pushes a new route", (tester) async {
-        final route = MaterialPageRoute<Object>(builder: (_) => Container());
-        await tester.pumpMaterialApp(
-          Builder(
-            builder: (context) => ElevatedButton(
-              onPressed: () async => context.push(route),
-              child: const Text("Push route"),
-            ),
-          ),
-        );
-        await tester.tap(find.byType(ElevatedButton));
-        await tester.pumpAndSettle();
-        final context = tester.element(find.byType(Container));
-        expect(find.byType(Container), findsOneWidget);
-        context.pop();
-        await tester.pumpAndSettle();
-        expect(find.byType(Container), findsNothing);
-      });
-
-      testWidgets(
-        "maybePop pops the current route if there is a previous route",
-        (tester) async {
-          final route1 = MaterialPageRoute<Object>(builder: (_) => Container());
-          final route2 =
-              MaterialPageRoute<Object>(builder: (_) => const SizedBox());
-          await tester.pumpMaterialApp(
-            Builder(
-              builder: (context) => ElevatedButton(
-                onPressed: () async {
-                  await context.push(route1);
-                  await context.push(route2);
-                },
-                child: const Text("Push routes"),
-              ),
-            ),
-          );
-          await tester.tap(find.byType(ElevatedButton));
-          await tester.pumpAndSettle();
-          var context = tester.element(find.byType(Container));
-          expect(find.byType(Container), findsOneWidget);
-          final result = await context.maybePop();
-          expect(result, isTrue);
-          await tester.pumpAndSettle();
-          expect(find.byType(SizedBox), findsOneWidget);
-          context = tester.element(find.byType(SizedBox))..pop();
-          await tester.pumpAndSettle();
-          expect(find.byType(Container), findsNothing);
-        },
-      );
-
-      testWidgets(
-        "maybePop",
-        (tester) async {
-          final context = await contextExtractor(tester);
-          final result = await context.maybePop();
-          expect(result, isFalse);
-        },
-      );
-
       testWidgets("focus", (tester) async {
         final context = await contextExtractor(tester);
         expect(context.focus, isA<FocusScopeNode>());
