+++ packages/world_countries/test/src/extensions/navigator_context_extension_test.dart
@@ -1,0 +1,97 @@
+import "package:flutter/material.dart";
+import "package:flutter_test/flutter_test.dart";
+import "package:world_countries/src/extensions/navigator_context_extension.dart";
+
+import "../../helpers/widget_tester_extension.dart";
+
+void main() => group("NavigatorContextExtension", () {
+      Future<BuildContext> contextExtractor(
+        WidgetTester tester, [
+        Widget child = const SizedBox(),
+      ]) async {
+        await tester.pumpMaterialApp(child);
+
+        return tester.element(find.byType(child.runtimeType));
+      }
+
+      testWidgets("navigator", (tester) async {
+        final context = await contextExtractor(tester);
+        expect(context.navigator, isA<NavigatorState>());
+        expect(context.navigator.canPop(), isFalse);
+      });
+
+      testWidgets("pop pops the current route", (tester) async {
+        final route = MaterialPageRoute<Object>(builder: (_) => Container());
+        await tester.pumpMaterialApp(
+          Builder(
+            builder: (context) => ElevatedButton(
+              onPressed: () async => context.push(route),
+              child: const Text("Push route"),
+            ),
+          ),
+        );
+        await tester.tapAndSettle(find.byType(ElevatedButton));
+        final context = tester.element(find.byType(Container));
+        expect(find.byType(Container), findsOneWidget);
+        context.pop();
+        await tester.pumpAndSettle();
+        expect(find.byType(Container), findsNothing);
+      });
+
+      testWidgets("push pushes a new route", (tester) async {
+        final route = MaterialPageRoute<Object>(builder: (_) => Container());
+        await tester.pumpMaterialApp(
+          Builder(
+            builder: (context) => ElevatedButton(
+              onPressed: () async => context.push(route),
+              child: const Text("Push route"),
+            ),
+          ),
+        );
+        await tester.tapAndSettle(find.byType(ElevatedButton));
+        final context = tester.element(find.byType(Container));
+        expect(find.byType(Container), findsOneWidget);
+        context.pop();
+        await tester.pumpAndSettle();
+        expect(find.byType(Container), findsNothing);
+      });
+
+      testWidgets(
+        "maybePop pops the current route if there is a previous route",
+        (tester) async {
+          final route1 = MaterialPageRoute<Object>(builder: (_) => Container());
+          final route2 =
+              MaterialPageRoute<Object>(builder: (_) => const SizedBox());
+          await tester.pumpMaterialApp(
+            Builder(
+              builder: (context) => ElevatedButton(
+                onPressed: () async {
+                  await context.push(route1);
+                  await context.push(route2);
+                },
+                child: const Text("Push routes"),
+              ),
+            ),
+          );
+          await tester.tapAndSettle(find.byType(ElevatedButton));
+          var context = tester.element(find.byType(Container));
+          expect(find.byType(Container), findsOneWidget);
+          final result = await context.maybePop();
+          expect(result, isTrue);
+          await tester.pumpAndSettle();
+          expect(find.byType(SizedBox), findsOneWidget);
+          context = tester.element(find.byType(SizedBox))..pop();
+          await tester.pumpAndSettle();
+          expect(find.byType(Container), findsNothing);
+        },
+      );
+
+      testWidgets(
+        "maybePop",
+        (tester) async {
+          final context = await contextExtractor(tester);
+          final result = await context.maybePop();
+          expect(result, isFalse);
+        },
+      );
+    });
